add LaTeX code to PNG metadata
using morgify to add the latex code used to generate the png to the
comments field of the png's metadata. this could potentially be useful
in the future to allow the image to be edited.

need to handle special escape sequences, for example the \r in \rowcolor
may be getting interpretted as a return character (not sure).